+++ src/collective/contact/core/tests/test_search.py
@@ -8,7 +8,7 @@ from ecreall.helpers.testing.base import BaseTest
 
 from collective.contact.core.testing import INTEGRATION
 from collective.contact.core.indexers import held_position_searchable_text,\
-    organization_searchable_text
+    organization_searchable_text, person_sortable_title
 
 
 class TestSearch(unittest.TestCase, BaseTest):
@@ -43,6 +43,12 @@ class TestSearch(unittest.TestCase, BaseTest):
         sergent_pepper = self.sergent_pepper
         self.assertEqual(held_position_searchable_text(sergent_pepper)(),\
                          u"Sergent Pepper Sergent de la brigade LH Armée de terre Corps A Division Alpha Régiment H Brigade LH")
+        pepper = self.pepper
+        self.assertEqual(person_sortable_title(pepper)(),
+                         u"Pepper")
+        degaulle = self.degaulle
+        self.assertEqual(person_sortable_title(degaulle)(),
+                         u"De Gaulle Charles")
 
     def test_searchable_fields(self):
         catalog = getToolByName(self.portal, 'portal_catalog')